+++ WebServerWithRestApi.App/PropertyNotFoundException.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Runtime.Serialization;
+
+namespace WebServerWithRestApi.App
+{
+    [Serializable]
+    internal class PropertyNotFoundException : Exception
+    {
+        public PropertyNotFoundException()
+        {
+        }
+
+        public PropertyNotFoundException(string message) : base(message)
+        {
+        }
+
+        public PropertyNotFoundException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        protected PropertyNotFoundException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}
