+++ WebServerWithRestApi.App/WebAppConfiguration.cs
@@ -6,7 +6,8 @@ namespace WebServerWithRestApi.App
     {
         public Dictionary<string, string> WebApps { get; private set; } = new Dictionary<string, string>()
         {
-            {"todo", @"C:\Users\rparkar\Desktop\Assignments\WebServerWithRestApi.App\WebApps\ToDo" }
+            {"todo", @"C:\Users\rparkar\Desktop\Assignments\WebServerWithRestApi.App\WebApps\ToDo" },
+            {"Error", @"C:\Users\rparkar\Desktop\Assignments\WebServerWithRestApi.App\WebApps\Error" }
         };
     }
 }
